+++ keyCommands.js
@@ -11,17 +11,16 @@
 const keyCommands = {
     'push_valider':  push_valider,
     'push_annuler': push_annuler,
-    // 'download_document': startDownload, => en attente de https://github.com/Refhi/Weda-Helper/pull/106
     'print_meds': function () {
-            getOption('RemoveLocalCompanionPrint', function (RemoveLocalCompanionPrint) {
-                if (!RemoveLocalCompanionPrint) {
-                    startPrinting('companion', 0);
-                } else {
-                    startPrinting('print', 0);
-                }
-            });
+            printIfOption(0);
+        },    
+    'print_meds_bis': function () {
+            printIfOption(1);
         },
     'download_document': function () {
+            startPrinting('download', 0);
+        },
+    'download_document_bis': function () {
             startPrinting('download', 1);
         },
     'upload_latest_file': uploadLatest,
@@ -121,17 +120,27 @@ function addShortcuts(keyCommands, scope, scopeName) {
 function addShortcutsToIframe() {
     var iframes = document.querySelectorAll('iframe');
     if (iframes.length !== 0) {
-        hotkeys.setScope('iframe1');
         iframes.forEach(function(iframe, index) {
+            let scopeName = 'iframe' + (index + 1);
+            hotkeys.setScope(scopeName);    
             console.log('iframe' + (index + 1), iframe);
-            addShortcuts(keyCommands, iframe.contentDocument, 'iframe' + (index + 1));
+            addShortcuts(keyCommands, iframe.contentDocument, scopeName);
         });
     }
 }
 
+function addAllShortcuts() {
+    hotkeys.unbind(); // nécessaire pour éviter les doublons de raccourcis clavier entrainant des doublons de documents...
+    addShortcuts(keyCommands, document, 'all');
+    addShortcutsToIframe();
+}
+
 // Ajout des raccourcis claviers sur le document racine
-addShortcuts(keyCommands, document, 'all');
-afterMutations(300, addShortcutsToIframe, "ajout raccourcis aux iframes"); // ajoute les raccourcis à toutes les iframes après chaque mutation du document
+setTimeout(function() {
+    addAllShortcuts();
+}, 20);
+afterMutations(300, addAllShortcuts, "ajout raccourcis aux iframes"); // ajoute les raccourcis à toutes les iframes après chaque mutation du document
+// ne pas mettre moins de 300ms sinon les raccourcis s'ajoutent quand même de façon cumulative
 
 
 
@@ -186,6 +195,7 @@ function push_valider() {
     }
     // click other elements, one after the other, until one of them works
     const actions = [
+        () => clickElementById('ButtonValidFileStream'),
         () => clickElementById('targetValider'), // utilisé quand j'ajoute une cible à un bouton
         () => clickElementById('ContentPlaceHolder1_BaseGlossaireUCForm1_ButtonValidDocument'),
         () => clickElementById('ContentPlaceHolder1_ButtonLibreValid'),
@@ -213,6 +223,17 @@ function push_annuler() {
     ];
 
     actions.some(action => action() !== false);
+}
+
+// Fonction permettant d'imprimer selon les options choisies
+function printIfOption(modelNumber = 0) {    
+    getOption('RemoveLocalCompanionPrint', function (RemoveLocalCompanionPrint) {
+        if (!RemoveLocalCompanionPrint) {
+            startPrinting('companion', modelNumber);
+        } else {
+            startPrinting('print', modelNumber);
+        }
+    });
 }
 
 // Définition de la fonction startPrinting