+++ update.js
@@ -6,31 +6,19 @@ function htmlMaker(text) {
 }
 
 var nouveautes = `
-# [2.3] - 2024-05-28
+# [2.4] - 28/06/2024
+## ajout :
+- Décocher automatiquement le message et le fichier IHE_XDM.zip lors de l'importation d'un message depuis la messagerie sécurisée
+- Sélection automatique du type d'ordonnance numérique quand il s'agit d'une Demande, et qu'un mot-clé est détecté : (infirmierRegex = /IDE|infirmier|pansement|injection/i; kineRegex = /kiné|kine|kinésithérapie|kinesitherapie|MKDE|kinesitherapeute|kinesithérapeute/i; pedicureRegex = /pédicure|pedicure|podologie|podologique|podologue/i; orthophonieRegex = /orthophonie|orthophonique|orthophoniste/i; let orthoptieRegex = /orthoptie|orthoptique|orthoptiste/i;) => n'hésitez pas à nous demander d'ajouter d'autres mot-clés pertinents.
+- 2 raccourcis désormais possibles selon le modèle d'impression preféré, idem pour les téléchargements : Ctrl+P pour l'impression et Ctrl+D pour le téléchargement pour le premier modèle, et Ctrl+Shift+P et Ctrl+Shift+D pour le second modèle.
+- détection automatique du Companion s'il est en route, mais n'est pas activé dans les options de Weda-Helper
 
-<strong>🔍 Peu de modifications visibles pour vous</strong>, mais beaucoup de travail en coulisses pour améliorer la stabilité et la maintenabilité du code.
-💡 Le plus gros changement est la refonte de la gestion des impressions et des téléchargements, qui devrait être plus fiable et plus rapide.
-⚙️ Les raccourcis claviers sont désormais directement gérés dans les options ! Vous devrez donc les redéfinir si vous les aviez personnalisés.
-
-
-## refactory :
-✅ passage des options par défaut dans le manifest.json pour éviter les doublons
-✅ creation et utilisation prioritaire de 'addTweak' qui simplifie l'ajout de fonctionnalités dans telle ou telle page en fonction de l'option liée. Pour faciliter la lecture du code et la maintenance on l'appelle après chaque ensemble de tableau urls/options/callbacks
-✅ refactory complet de la gestion des impressions et des téléchargements
-✅ Les raccourcis claviers sont désormais directement gérés dans les options ! Vous devrez donc les redéfinir si vous les aviez personnalisés.
-
-## ajout :
-🆕 on peut désormais uploader un document en un seul raccourcis clavier ! (par défaut Ctrl+U) Définissez le dossier dans le companion (v1.4+). Ctrl+U enverra automatiquement le dernier fichier créé. (nécessite le Companion v1.4+)
-🆕 création d'une fiche-mémo <a href="https://github.com/Refhi/Weda-Helper/releases/latest/download/FicheMemo.pdf" target="_blank">disponible ici</a> pour vous aider à vous familiariser avec les raccourcis claviers et les fonctionnalités de Weda-Helper
-
-# Companion v1.4.2 !
-🆕 ajout de la possibilité de définir le dossier d'upload automatique
-🆕 ajout d'une interface graphique
-🆕 retrait de la console noire qui s'ouvrait à chaque lancement, remplacée par un "W" dans la barre des tâches
-🆕 une version Mac !
-=> par ici pour les détails et les téléchargements : <a href="https://github.com/Refhi/Weda-Helper-Companion" target="_blank">Weda-Helper Companion</a>
-🆕 installation très grandement simplifiée (récupération clé API automatique, réglages par défaut immédiatements fonctionnels, etc.)
-
+## fix :
+- le focus reste dans le champ de recherche après l'ouverture automatique des prescriptions-types
+- blocage du historyToLeft si une fenêtre de prévisualisation est ouverte pour éviter des superpositions
+- alt+V fonctionne désormais aussi pour valider l'import de documents
+- message [addTweak] plus explicite dans la console
+- amélioration du Readme, notamment la partie sur le Companion et les raccourcis claviers
 `
 
 nouveautes = htmlMaker(nouveautes)
@@ -47,6 +35,8 @@ var firstStartMessage = `
 🔧 Je vous encourage également à installer le Companion <a href="https://github.com/Refhi/Weda-Helper-Companion/releases/latest/download/Weda.Companion.exe" target="_blank">disponible ici pour windows</a> et <a href="https://github.com/Refhi/Weda-Helper-Companion/releases/latest/download/Weda.Companion.dmg" target="_blank">ou pour mac</a> pour profiter de fonctionnalités supplémentaires (Impression totale, lien avec le TPE et upload automatisé).
 
 📝 Vous pouvez aussi relire <a href="https://github.com/Refhi/Weda-Helper/" target="_blank">Weda-Helper sur gitHub</a> pour plus de précisions, et y faire des suggestions ou des signalements de bugs. 
+
+<a href="https://github.com/Refhi/Weda-Helper/blob/main/FicheMemo.pdf" target="_blank">📄 Fiche mémo</a> pour les raccourcis clavier.
 
 💖 Et bien sûr m'encourager sur le <a href="https://communaute.weda.fr/t5/Entraide-Logiciel-Weda/Weda-Helper-et-Weda-Helper-Companion/m-p/2998" target="_blank">Site de la communauté de weda</a>
 
@@ -70,6 +60,8 @@ var updateMessage = `
 🔧 Je vous conseille d'aller faire un tour dans les options pour vérifier les nouveaux paramètres : bouton de droite sur l'icone de l'extension puis option.
 
 🚀 Si vous ne l'avez pas encore, n'hésitez pas à tester le Companion : <a href="https://github.com/Refhi/Weda-Helper-Companion/releases/latest/download/Weda.Companion.exe" target="_blank">disponible ici pour windows</a> et <a href="https://github.com/Refhi/Weda-Helper-Companion/releases/latest/download/Weda.Companion.dmg" target="_blank">ou pour mac</a> pour profiter de fonctionnalités supplémentaires (Impression totale, lien avec le TPE et upload automatisé).</strong>
+
+<a href="https://github.com/Refhi/Weda-Helper/blob/main/FicheMemo.pdf" target="_blank">📄 et à télécharger la Fiche mémo</a> pour les raccourcis clavier.
 
 <strong>🌟 Voici les nouveautés et les améliorations :</strong>
 ${nouveautes}